backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
sqlalchemy
asyncpg
deepgram-sdk
openai
pydantic
pydantic-settings
passlib[bcrypt]
PyJWT
